+++ CHANGELOG.md
@@ -17,6 +17,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 * Vector sample data for local GeoServer (sea mask, coastline, graticule and facilities)
 * Option for using un-minified JS/CSS for debugging scripts
 * Minimal CI/CD pipeline to publish app WAR file as an artefact
+* Default support for the field parties plugin
 
 ### Changed